pyproject.toml
[tool.poetry]
name = "bepro_back"
version = "0.1.0"
description = ""
authors = ["Bakhtiyar Ospanov <b.p.ospanov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
django-filter = "^21.1"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.3"
django-environ = "^0.8.1"
djangorestframework-simplejwt = "^5.2.0"
djangorestframework-jsonapi = "^5.0.0"
Pillow = "^9.1.1"
django-db-logger = "^0.1.12"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
